uds2: move merge_nix_configs to shared
